Update vector-normalize-transact-sql.md
I don't know too much about vectors, but the sentence as given does not make much sense. But I can see that if I use norm1, the sum of the absolute value of the elements is 1. If I use norm2, the sum of the squares is 1. And with norminf, the biggest absolute value is 1.

Someone who knows more about vectors maybe could elaborate more on the length. (Which I first read as "number of elements", which is the more computeresque interpretation of what is a mathematical term here.)